Refactor common utilities
