+++ src/tennis_cut/tennis_cut.py
@@ -9,12 +9,15 @@ import argparse
 import json
 import logging
 import shutil
-import sys
 import tempfile
 import warnings
 from dataclasses import dataclass
 from pathlib import Path
-from typing import List, Sequence, Tuple
+from typing import List, Sequence
+
+import sys
+sys.path.append(str(Path(__file__).resolve().parents[1]))
+from utilities import PersonDetector, expand_box
 
 import subprocess
 import torchaudio
@@ -132,31 +135,6 @@ class PopDetector:
         return peaks
 
 
-class PersonDetector:
-    """Wrapper around YOLOv8-n person detector."""
-
-    def __init__(self, device: str) -> None:
-        from ultralytics import YOLO
-
-        self.device = device
-        self.model = YOLO("yolov8n.pt")
-
-    def find_box(self, img_path: Path) -> Tuple[int, int, int, int] | None:
-        """Return the largest person box as (x1, y1, x2, y2) integers."""
-
-        results = self.model.predict(
-            str(img_path), classes=0, device=self.device, verbose=False
-        )
-        boxes = results[0].boxes
-        if boxes is None or boxes.xyxy.numel() == 0:
-            return None
-        xyxy = boxes.xyxy.cpu().numpy()
-        areas = (xyxy[:, 2] - xyxy[:, 0]) * (xyxy[:, 3] - xyxy[:, 1])
-        idx = areas.argmax()
-        x1, y1, x2, y2 = xyxy[idx]
-        return int(x1), int(y1), int(x2), int(y2)
-
-
 def extract_frame(video: Path, time: float, out_path: Path) -> None:
     """Extract a single frame from *video* at *time* seconds."""
 
@@ -171,45 +149,6 @@ def extract_frame(video: Path, time: float, out_path: Path) -> None:
         str(out_path),
         "-y",
     ])
-
-
-def expand_box(box: Tuple[int, int, int, int], res: Tuple[int, int]) -> Tuple[int, int, int, int]:
-    """Pad box by 20% and fit to the video aspect ratio."""
-
-    x1, y1, x2, y2 = box
-    w = x2 - x1
-    h = y2 - y1
-    cx = x1 + w / 2
-    cy = y1 + h / 2
-    w *= 1.2
-    h *= 1.2
-    frame_w, frame_h = res
-    aspect = frame_w / frame_h
-    if w / h > aspect:
-        h = w / aspect
-    else:
-        w = h * aspect
-
-    # Ensure the crop doesn't exceed the frame size
-    if w > frame_w:
-        w = frame_w
-    if h > frame_h:
-        h = frame_h
-
-    x1 = int(round(cx - w / 2))
-    y1 = int(round(cy - h / 2))
-    w = int(round(w))
-    h = int(round(h))
-
-    if x1 < 0:
-        x1 = 0
-    if y1 < 0:
-        y1 = 0
-    if x1 + w > frame_w:
-        x1 = max(0, frame_w - w)
-    if y1 + h > frame_h:
-        y1 = max(0, frame_h - h)
-    return x1, y1, w, h
 
 
 def parse_args(argv: Sequence[str]) -> argparse.Namespace:
