+++ src/train_swing_detector/prepare_swing_frames.py
@@ -21,87 +21,13 @@ import random
 import subprocess
 import sys
 from tqdm import tqdm
-from ultralytics import YOLO
+sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
+from utilities import PersonDetector, expand_box, probe_duration
 
 # -----------------------  constants / defaults  -----------------------
 NEG_PER_POS = 1            # number of no_shot samples per positive
 MIN_OFF     = 1            # negatives are at least this far from impact
 MAX_OFF     = 2            # and at most this far from impact
-
-class PersonDetector:
-    """Wrapper around YOLOv8-n person detector."""
-
-    def __init__(self, device: str) -> None:
-        self.device = device
-        self.model = YOLO("yolov8n.pt")
-
-    def find_box(self, img_path: pathlib.Path) -> tuple[int, int, int, int] | None:
-        """Return the largest person box as (x1, y1, x2, y2) integers."""
-        results = self.model.predict(
-            str(img_path), classes=0, device=self.device, verbose=False
-        )
-        boxes = results[0].boxes
-        if boxes is None or boxes.xyxy.numel() == 0:
-            return None
-        xyxy = boxes.xyxy.cpu().numpy()
-        areas = (xyxy[:, 2] - xyxy[:, 0]) * (xyxy[:, 3] - xyxy[:, 1])
-        idx = areas.argmax()
-        x1, y1, x2, y2 = xyxy[idx]
-        return int(x1), int(y1), int(x2), int(y2)
-
-
-def expand_box(box: tuple[int, int, int, int], res: tuple[int, int]) -> tuple[int, int, int, int]:
-    """Pad box by 20% and fit to the video aspect ratio."""
-    x1, y1, x2, y2 = box
-    w = x2 - x1
-    h = y2 - y1
-    cx = x1 + w / 2
-    cy = y1 + h / 2
-    w *= 1.2
-    h *= 1.2
-    frame_w, frame_h = res
-    aspect = frame_w / frame_h
-    if w / h > aspect:
-        h = w / aspect
-    else:
-        w = h * aspect
-
-    # Ensure the crop doesn't exceed the frame size
-    if w > frame_w:
-        w = frame_w
-    if h > frame_h:
-        h = frame_h
-
-    x1 = int(round(cx - w / 2))
-    y1 = int(round(cy - h / 2))
-    w = int(round(w))
-    h = int(round(h))
-
-    if x1 < 0:
-        x1 = 0
-    if y1 < 0:
-        y1 = 0
-    if x1 + w > frame_w:
-        x1 = max(0, frame_w - w)
-    if y1 + h > frame_h:
-        y1 = max(0, frame_h - h)
-    return x1, y1, w, h
-
-
-def probe_duration(path: pathlib.Path) -> float:
-    """Return media duration in seconds using ffprobe."""
-    cmd = [
-        "ffprobe",
-        "-v",
-        "error",
-        "-show_entries",
-        "format=duration",
-        "-of",
-        "default=noprint_wrappers=1:nokey=1",
-        str(path),
-    ]
-    out = subprocess.check_output(cmd).decode().strip()
-    return float(out)
 
 
 def extract_frame(mp4_path: pathlib.Path, timestamp: float, out_path: pathlib.Path, detector: PersonDetector, resolution: tuple[int, int]) -> None:
@@ -123,7 +49,7 @@ def extract_frame(mp4_path: pathlib.Path, timestamp: float, out_path: pathlib.Pa
         "-y",
     ]
     try:
-        result = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True, text=True)
+        subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True, text=True)
     except subprocess.CalledProcessError as e:
         print(f"Error during ffmpeg execution: {e.stderr.strip()}")
         raise
@@ -142,7 +68,7 @@ def extract_frame(mp4_path: pathlib.Path, timestamp: float, out_path: pathlib.Pa
             "-y",
         ]
         try:
-            result = subprocess.run(crop_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True, text=True)
+            subprocess.run(crop_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True, text=True)
         except subprocess.CalledProcessError as e:
             print(f"Error during ffmpeg crop execution: {e.stderr.strip()}")
             raise
